Re-named and re-organized directories to better handle future components. Slight tweaks and edits to styling and layout for h1's and about me/desc sections.
