+++ src/components/posts/GearDetails.css
@@ -44,7 +44,7 @@
 
 .gear-description p {
   margin-top: 10px;
-  text-indent: 20px;
+  white-space: pre-wrap;
 }
 
 .button-container {
