+++ src/components/posts/LargePosts.css
@@ -61,6 +61,7 @@
     margin: 15px 0; /* Add space around the description */
     padding-left: 20px; /* Indent the description from the left */
     color: #003231;
+    white-space: pre-wrap;
 }
 
 .large-post-link {
