+++ src/components/posts/SmallPosts.css
@@ -1,5 +1,3 @@
-/* SMALL POST - CUSTOM STYLES */
-
 .small-card {
     box-shadow: 0px 0px 6px 4px rgba(205, 214, 187, 0.5);
     border: none;
